[chaging_default_domain] Extracting Meta class from manifest #255
